refactor: split handlers into single files
