+++ src/handlers.rs
@@ -1,1 +1,3 @@
+pub mod play_sound;
 pub mod sounds;
+pub mod upload;
